move handleDependencies into its own module
